# openapi.yaml
openapi: 3.1.0
info:
  title: CKPool HTTP Bridge
  version: 1.1.0
  description: Persistent HTTP/JSON facade over CKPool Stratum V1, with an auto-reconnecting default worker.
servers:
  - url: https://YOUR-BRIDGE-HOST
paths:
  /health:
    get:
      operationId: health
      responses:
        '200': { description: Bridge health and managed-session status }
  /v1/default:
    get:
      operationId: getDefaultSession
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Auto-reconnecting default CKPool session }
  /v1/default/events:
    get:
      operationId: getDefaultEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: after
          schema: { type: integer, default: 0 }
        - in: query
          name: wait
          schema: { type: integer, minimum: 0, maximum: 25000, default: 0 }
      responses:
        '200': { description: Managed Stratum events after sequence number }
  /v1/default/submit:
    post:
      operationId: submitDefaultShare
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Share'
      responses:
        '200': { description: CKPool share acceptance result }
  /v1/default/reconnect:
    post:
      operationId: reconnectDefaultSession
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: wait
          schema: { type: integer, minimum: 0, maximum: 25000, default: 12000 }
      responses:
        '200': { description: Reconnect requested and current managed status }
  /v1/session:
    post:
      operationId: createSession
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address: { type: string }
                worker: { type: string }
                password: { type: string, default: x }
                jobWaitMs: { type: integer, default: 10000 }
      responses:
        '201': { description: Connected Stratum session }
  /v1/session/{sessionId}:
    get:
      operationId: getSession
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Session status and current job }
    delete:
      operationId: closeSession
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Session closed }
  /v1/session/{sessionId}/events:
    get:
      operationId: getEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
        - in: query
          name: after
          schema: { type: integer, default: 0 }
        - in: query
          name: wait
          schema: { type: integer, minimum: 0, maximum: 25000, default: 0 }
      responses:
        '200': { description: Stratum events after sequence number }
  /v1/session/{sessionId}/submit:
    post:
      operationId: submitShare
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Share'
      responses:
        '200': { description: Pool acceptance result }
components:
  schemas:
    Share:
      type: object
      required: [jobId, extranonce2, ntime, nonce]
      properties:
        jobId: { type: string }
        extranonce2: { type: string }
        ntime: { type: string, pattern: '^[0-9A-Fa-f]{8}$' }
        nonce: { type: string, pattern: '^[0-9A-Fa-f]{8}$' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer